+++ server/pkg/macAPI/display.go
@@ -1,4 +1,4 @@
-package api
+package macAPI
 
 import (
 	"github.com/andybrewer/mack"
