+++ server/pkg/macAPI/keyboard.go
@@ -1,4 +1,4 @@
-package api
+package macAPI
 
 import (
 	"fmt"
